Add voters in the code

+++ app/src/Controller/Front/PropertyController.php
@@ -12,6 +12,7 @@ use App\Entity\Property;
 use App\Form\OwnedPropertiesSearchType;
 use App\Form\PropertyType;
 use App\Repository\UserRepository;
+use App\Security\Voter\OwnedPropertyVoter;
 use Doctrine\ORM\EntityManagerInterface;
 use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
@@ -37,6 +38,10 @@ class PropertyController extends AbstractController
             $paginator = $propertyRepository->getPropertiesPaginationForUser($this->getUser(), $request, 4);
         }
 
+        foreach ($paginator as $property) {
+            $this->denyAccessUnlessGranted(OwnedPropertyVoter::VIEW, $property);
+        }
+
         return $this->render('front/property/index.html.twig', [
             'paginator' => $paginator,
             'searchForm' => $form->createView()
@@ -51,6 +56,8 @@ class PropertyController extends AbstractController
      */
     public function new(Request $request, SearchRepository $searchRepository): Response
     {
+        $this->denyAccessUnlessGranted(OwnedPropertyVoter::CREATE, Property::class);
+
         $property = new Property();
         $property->setOwner($this->getUser());
         $form = $this->createForm(PropertyType::class, $property);
@@ -76,22 +83,12 @@ class PropertyController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="property_show", methods={"GET"})
-     * @param Property $property
-     * @return Response
-     */
-    public function show(Property $property): Response
-    {
-        return $this->render('front/property/show.html.twig', [
-            'property' => $property,
-        ]);
-    }
-
-    /**
      * @Route("/{id}/edit", name="property_edit", methods={"GET","POST"})
      */
     public function edit(Request $request, Property $property): Response
     {
+        $this->denyAccessUnlessGranted(OwnedPropertyVoter::UPDATE, $property);
+
         $form = $this->createForm(PropertyType::class, $property);
         $form->handleRequest($request);
 
@@ -113,6 +110,8 @@ class PropertyController extends AbstractController
      */
     public function delete(Request $request, Property $property): Response
     {
+        $this->denyAccessUnlessGranted(OwnedPropertyVoter::DELETE, $property);
+
         if ($this->isCsrfTokenValid('delete' . $property->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($property);
